requirements.txt
-i https://pypi.org/simple
pillow==12.1.1; python_version >= '3.10'
pyside6==6.10.2; python_version < '3.15' and python_version >= '3.9'
pyside6-addons==6.10.2; python_version < '3.15' and python_version >= '3.9'
pyside6-essentials==6.10.2; python_version < '3.15' and python_version >= '3.9'
shiboken6==6.10.2; python_version < '3.15' and python_version >= '3.9'
